DTMF dialpad pops up automatically after changing language
When user changes language, the InCallActivity is destroyed and then
re-created. After InCallActivity is re-created, the dial pad fragment
is restored by the framework, and dial pad view is re-created. So the
dial pad becomes visible.

Set the default visibility of dial pad view to gone, so even if the
framework restores the dial pad fragment, the view is still invisible.
And we also need to uncheck the dial pad button as the dial pad view
is invisible.

Change-Id: Idfc3e9e4ec5240809b77f77108d048d357e15b91